Added sigint handler to server.cpp

+++ src/player.h
@@ -15,9 +15,9 @@ public:
 	std::string username = "player";
     Map* map;
 
-	int xp = 0;
-	int reputation = 0;
-	int currency = 0;
+	unsigned short xp = 0;
+	unsigned short reputation = 0;
+	unsigned short currency = 0;
 
 	void give_xp(int amount);
 	int xp_to_level();
